requirements.txt
selenium>=4.22.0
python-dotenv>=1.0.1
pyyaml>=6.0.2
